Further help for PEP 8, and disable the buttons until file is open
Kyzz and I were working on this on #openhatch today, basically I
put together a new tiny feature that was pretty helpful as a
result. Also, the README is updated with cool information about
Fedora, and there is some good work happening everywhere.

Oh, and whitespace fixes. Don't know how that happened.